requirements.txt
# Core ML dependencies
torch
transformers
tokenizers
datasets
numpy
scikit-learn

# Web app and API
streamlit
requests
newsapi-python

# Configuration and environment
python-dotenv
PyYAML

# Visualization and monitoring
matplotlib
seaborn
tqdm

# Utilities
gdown

